eliminating UB
Shifting 32-bit value by 48 bits is undefined behaviour

+++ code/components/loading-screens-five/src/LoadingScreens.cpp
@@ -776,7 +776,7 @@ static InitFunction initFunction([] ()
 				}
 			}
 
-			uint64_t loadTimingIdentity = HashString(name) | (16 << 48);
+			uint64_t loadTimingIdentity = HashString(name) | (uint64_t(16) << 48);
 
 			UpdateLoadTiming(loadTimingIdentity);
